# openapi.yaml
openapi: 3.0.3
info:
  title: API Potencial Efficopar
  description: |
    API para cálculo de potencial e análise de risco de inadimplência e medição.
    
    Esta API fornece endpoints para:
    - Gerenciamento de grupos e setores
    - Consulta de parâmetros de risco
    - Cálculo de ranges e pesos automatizados
    - Catálogo de entidades disponíveis
  version: 1.0.0
  contact:
    name: Aceno
    url: https://aceno.com.br
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api-potencial.efficopar.com.br
    description: Servidor de produção
...
paths:
  /grupos:
    post:
      tags:
        - Grupos
      summary: Criar ou atualizar grupo de setores
      description: |
        Cria um novo grupo de setores ou atualiza um existente caso o nome já exista.
        O grupo permite agrupar múltiplos setores para análise conjunta.
      operationId: createOrUpdateGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - setores
              properties:
                nome:
                  type: string
                  description: Nome do grupo (se não informado, será gerado automaticamente)
                  example: "Grupo Central"
                setores:
                  type: array
                  description: Lista de códigos dos setores que compõem o grupo
                  minItems: 1
                  items:
                    type: string
                  example: ["SETOR01", "SETOR02", "SETOR03"]
            examples:
              comNome:
                summary: Criar grupo com nome específico
                value:
                  nome: "Grupo Industrial"
                  setores: ["IND001", "IND002", "IND003"]
              semNome:
                summary: Criar grupo com nome automático
                value:
                  setores: ["COM001", "COM002"]
      responses:
        '200':
          description: Grupo criado ou atualizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  grupo_id:
                    type: string
                    format: uuid
                    description: ID do grupo criado/atualizado
                    example: "550e8400-e29b-41d4-a716-446655440000"
                  message:
                    type: string
                    example: "Grupo criado/atualizado com sucesso"
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                setoresVazio:
                  value:
                    error: "setores must be a non-empty array"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: "Internal server error"
...